from uint8 to uint16 in i1 CSI-RS measurement fields

+++ nfapi/open-nFAPI/nfapi/public_inc/fapi_nr_ue_interface.h
@@ -54,7 +54,7 @@ typedef struct {
   uint32_t rsrp;
   int rsrp_dBm;
   uint8_t rank_indicator;
-  uint8_t i1;
+  uint16_t i1;
   uint8_t i2;
   uint8_t cqi;
   rlm_t radiolink_monitoring;
